requirements.txt
pillow
py-cord
python-dotenv
